Fix Mac compatibility issues with test framework
This commit addresses cross-platform compatibility issues between Windows and Mac:
1. Modified GUT.cs to conditionally compile test methods and attributes
2. Added compiler constants to SignalLost.csproj to prevent duplicate definitions
3. Fixed type conversion issue in SignalStrengthTest.cs

These changes ensure the project builds and runs correctly on both platforms.

+++ godot_project/scripts/SignalStrengthTest.cs
@@ -40,7 +40,7 @@ namespace SignalLost
             _perfectButton.Pressed += () => SetSignalStrength(1.0f);
 
             // Set initial signal strength
-            SetSignalStrength(_slider.Value / 100.0f);
+            SetSignalStrength((float)_slider.Value / 100.0f);
 
             GD.Print("SignalStrengthTest ready!");
         }
